+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Integration Camera Scanner SDK
+##Added
+- implementation of new strategy to launch the scanner camera.
 # v2.8.2
 ## Fixed
 - Implement sending payment flow response data to PTM